feat: добавляет в Провайдер проверку на online

+++ src/api/provider.js
@@ -1,21 +1,45 @@
+const isOnline = () => {
+  return window.navigator.onLine;
+};
+
 export default class Provider {
   constructor(api) {
     this._api = api;
   }
 
   addEvent(data) {
-    return this._api.addEvent(data);
+    if (isOnline()) {
+      return this._api.addEvent(data);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   deleteEvent(id) {
-    return this._api.deleteEvent(id);
+    if (isOnline()) {
+      return this._api.deleteEvent(id);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   getData() {
-    return this._api.getData();
+    if (isOnline()) {
+      return this._api.getData();
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   updateEvent(id, data) {
-    return this._api.updateEvent(id, data);
+    if (isOnline()) {
+      return this._api.updateEvent(id, data);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 }
